app/requirements.txt
appnope==0.1.0
asgiref==3.3.0
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
decorators==2.0.7
dj-database-url==0.5.0
Django==3.1.3
django-autocomplete-light==3.5.1
django-bootstrap-modal-forms==2.0.0
django-crispy-forms==1.9.2
django-datatables-view==1.19.1
django-heroku==0.3.1
django-widget-tweaks==1.4.2
django-tempus-dominus==5.1.2.13
environs==8.0.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
marshmallow==3.9.0
openpyxl==3.0.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.2
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
Unipath==1.1
urllib3==1.25.11
wcwidth==0.2.5
whitenoise==5.2.0
